requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.5
transformers>=4.30.0
tokenizers>=0.13.3
pandas>=1.5.3
numpy>=1.23.5
scipy>=1.10.0
Pillow>=9.4.0
tqdm>=4.65.0
